feat: creating and joining board done
