+++ src/pages/api/board/create.ts
@@ -1,42 +1,13 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
-import { currentDate } from '@/utils/constants';
+import { createBoard } from '@/utils/api/board';
+import { User, addBoardtoFacilitator, createUser } from '@/utils/api/user';
 import { ResponseInterface } from '@/utils/interface';
-import { db } from '@lib/firebase/client';
-import { addDoc, collection } from 'firebase/firestore';
-import { customAlphabet, nanoid } from 'nanoid';
-import { alphanumeric } from 'nanoid-dictionary';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
 interface Data extends ResponseInterface {
   boardCode?: string;
   message?: string;
 }
-
-const createBoard = async (name: string, facilitator: string) => {
-  const generateInvite = customAlphabet(alphanumeric, 6);
-  const boardCode = generateInvite();
-
-  try {
-    const boardRef = collection(db, 'boards');
-    const newBoard = {
-      name,
-      boardId: nanoid(),
-      boardCode,
-      facilitator,
-      members: [facilitator],
-      allPosts: [],
-      createdAt: currentDate,
-    };
-    const docRef = await addDoc(boardRef, newBoard);
-
-    return { status: 'success', code: boardCode };
-  } catch (error) {
-    console.log({ error });
-    if (error instanceof Error) {
-      return { status: 'error', message: error.message };
-    }
-  }
-};
 
 export default async function handler(
   req: NextApiRequest,
@@ -45,16 +16,53 @@ export default async function handler(
   if (req.method !== 'POST') {
     return res.status(400).json({ status: 'error' });
   }
+  /**
+   * A Facilitator account is created, then a board is created and then updated with facilitator details
+   */
 
-  const { name, facilitator } = req.body;
-  const response = await createBoard(name, facilitator);
-  if (response?.status === 'success') {
-    return res
-      .status(200)
-      .json({ status: response.status, boardCode: response.code });
+  const {
+    userName: name,
+    interests,
+    profilePic,
+    areaOfResidence,
+    boardName,
+  } = req.body;
+
+  const userData: User = {
+    name,
+    interests,
+    profilePic,
+    areaOfResidence,
+  };
+
+  const create = await createUser(userData, 'facilitator');
+  const userIsCreated = create?.status === 'success';
+
+  if (userIsCreated && create?.userId) {
+    const facilitator = { name: name, id: create?.userId };
+    const response = await createBoard(boardName, facilitator);
+
+    if (response?.status === 'success') {
+      // if board is successfully creeted, update the board with user details
+      const boardToUser = await addBoardtoFacilitator(
+        create?.userId,
+        boardName
+      );
+
+      if (boardToUser?.status === 'success') {
+        return res
+          .status(200)
+          .json({ status: response.status, boardCode: response.code });
+      }
+    } else {
+      return res.status(400).json({
+        status: 'error',
+        message: response?.message ?? 'Error addding facilitator to board',
+      });
+    }
   } else {
     return res
       .status(400)
-      .json({ status: 'error', message: response?.message });
+      .json({ status: 'error', message: 'Error creating board' });
   }
 }
