+++ src/pages/api/board/join/[code].ts
@@ -1,17 +1,7 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
-import { currentDate } from '@/utils/constants';
+import { addUserToBoard, getBoardData } from '@/utils/api/board';
+import { User, createUser } from '@/utils/api/user';
 import { ResponseInterface } from '@/utils/interface';
-import { db } from '@lib/firebase/client';
-import {
-  addDoc,
-  arrayUnion,
-  collection,
-  getDocs,
-  query,
-  updateDoc,
-  where,
-} from 'firebase/firestore';
-import { nanoid } from 'nanoid';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
 interface Data extends ResponseInterface {
@@ -22,87 +12,6 @@ interface Data extends ResponseInterface {
     facilitator: string;
   };
 }
-
-type BoardType =
-  | {
-      name: string;
-      createdAt: string;
-      facilitator: string;
-      boardCode: string;
-      members: string[];
-      boardId: string;
-      allPosts: unknown[];
-    }
-  | undefined;
-
-type User = {
-  name: string;
-  interests: string[];
-  profilePic: string;
-  areaOfResidence: string;
-  groups?: string[];
-};
-
-/**
- * TODO : figure out a way to have a unique accessible data for traversing users
- * since we're not authenticating, we need a username/unique identifier
- *
- * 2. find out how to structure these functions
- *
- * 3. create facilitator account as they're creating board as well
- */
-
-async function getBoardData(boardCode: string): Promise<BoardType> {
-  const boardRef = collection(db, 'boards');
-  const q = query(boardRef, where('boardCode', '==', boardCode));
-  const snapshot = await getDocs(q);
-  let document;
-  if (snapshot.empty) {
-    console.log('No matching document');
-  } else {
-    snapshot.forEach((doc) => {
-      document = doc.data();
-    });
-  }
-  return document;
-}
-
-const createUser = async (user: User, role = 'member') => {
-  try {
-    const userRef = collection(db, 'users');
-    const userId = nanoid();
-    const newUser = {
-      userId,
-      name: user.name,
-      interests: user.interests,
-      profilePic: user.profilePic,
-      areaOfResidence: user.areaOfResidence,
-      boards: [],
-      groups: [],
-      allPosts: [],
-      role,
-      createdAt: currentDate,
-    };
-    const docRef = await addDoc(userRef, newUser);
-    return { status: 'success', userId };
-  } catch (error) {
-    console.log({ error });
-    if (error instanceof Error) {
-      return { status: 'error', message: error.message };
-    }
-  }
-};
-
-const addUserToBoard = async (userId: string, boardCode: string) => {
-  const boardsRef = collection(db, 'boards');
-  const q = query(boardsRef, where('boardCode', '==', boardCode));
-  const querySnapshot = await getDocs(q);
-
-  if (!querySnapshot.empty) {
-    const doc = querySnapshot.docs[0];
-    await updateDoc(doc.ref, { members: arrayUnion(userId) });
-  }
-};
 
 export default async function handler(
   req: NextApiRequest,
@@ -128,7 +37,7 @@ export default async function handler(
       .json({ status: 'success', data: { name, facilitator } });
   }
 
-  // assuming creating an account and joining a board are coupled
+  // Creates User Account and then joins board
   if (req.method === 'POST') {
     const { name: userName, interests, profilePic, areaOfResidence } = req.body;
     const userData: User = {
@@ -137,12 +46,13 @@ export default async function handler(
       profilePic,
       areaOfResidence,
     };
-    const response = await createUser(userData);
+    const response = await createUser(userData, 'member');
     console.log({ response });
 
     if (response?.userId) {
       try {
-        await addUserToBoard(response.userId, code);
+        const user = { id: response?.userId, name: userData.name };
+        await addUserToBoard(user, code);
         return res
           .status(200)
           .json({ status: 'success', message: 'User Added to Board' });
